Changed behavior so as not to indicate the error message box when DynamicVariables is 1 and FontSize is 0. (issue 126) FIXED: Virtual Bytes usage for an x64 process with perfmon plugin does not display value greater than 4 GB (issue 113) The numerical value greater than 32bit can be now displayed when NumOfDecimals is 0 in Meter=STRING. Fixed the broken compatibility in Windows 2000. (AboutDialog, SysInfo)

+++ Library/Measure.cpp
@@ -481,7 +481,7 @@ const WCHAR* CMeasure::GetStringValue(bool autoScale, double scale, int decimals
 		{
 			double val = theValue * (1.0 / scale);
 		    val = (val + ( (val >= 0) ? 0.5 : -0.5 ) );
-			swprintf(buffer, L"%i", (UINT)val);
+			swprintf(buffer, L"%lli", (LONGLONG)val);
 		}
 		else
 		{
